upgrade to use the new SHA3 implementation
to leverage the faster new interfaces update() and digest()
